Adding support for design tokens in DrawerController
